requirements.txt
fastapi==0.115.8
uvicorn==0.27.1
gunicorn==23.0.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.10
asyncpg==0.29.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.2.1
requests==2.32.5
httpx==0.27.0
groq==0.16.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.160.0
apscheduler==3.11.0
pytz==2024.1
slowapi==0.1.9
aiofiles==24.1.0
python-dateutil==2.9.0.post0
bcrypt==4.2.1
chromadb==0.6.3
sentence-transformers==5.2.0
email-validator==2.2.0
prometheus-fastapi-instrumentator>=7.0.0
celery==5.3.6
tenacity==8.2.3
redis>=5.0.0
cloudinary>=1.38.0
pytest-asyncio>=0.23.0
--extra-index-url https://download.pytorch.org/whl/cpu
aiosqlite>=0.19.0
psutil>=5.9.0
